move failed trp-ds tests to seperate file
