+++ internal/venice-test-common/src/integrationTest/java/com/linkedin/venice/endToEnd/TestDeferredVersionSwap.java
@@ -6,13 +6,11 @@ import static com.linkedin.venice.ConfigKeys.CONTROLLER_DEFERRED_VERSION_SWAP_SL
 import static com.linkedin.venice.ConfigKeys.DATA_BASE_PATH;
 import static com.linkedin.venice.ConfigKeys.LOCAL_REGION_NAME;
 import static com.linkedin.venice.utils.IntegrationTestPushUtils.createStoreForJob;
-import static com.linkedin.venice.utils.TestUtils.assertCommand;
 import static com.linkedin.venice.utils.TestWriteUtils.NAME_RECORD_V3_SCHEMA;
 import static com.linkedin.venice.utils.TestWriteUtils.getTempDataDirectory;
 import static com.linkedin.venice.vpj.VenicePushJobConstants.TARGETED_REGION_PUSH_LIST;
 import static com.linkedin.venice.vpj.VenicePushJobConstants.TARGETED_REGION_PUSH_WITH_DEFERRED_SWAP;
 import static org.testng.Assert.assertEquals;
-import static org.testng.Assert.assertFalse;
 import static org.testng.Assert.assertNotNull;
 import static org.testng.Assert.assertNull;
 import static org.testng.Assert.assertTrue;
@@ -28,11 +26,9 @@ import com.linkedin.venice.client.store.StatTrackingStoreClient;
 import com.linkedin.venice.common.VeniceSystemStoreType;
 import com.linkedin.venice.controllerapi.ControllerClient;
 import com.linkedin.venice.controllerapi.ControllerResponse;
-import com.linkedin.venice.controllerapi.JobStatusQueryResponse;
 import com.linkedin.venice.controllerapi.StoreResponse;
 import com.linkedin.venice.controllerapi.UpdateStoreQueryParams;
 import com.linkedin.venice.controllerapi.VersionCreationResponse;
-import com.linkedin.venice.exceptions.VeniceException;
 import com.linkedin.venice.integration.utils.D2TestUtils;
 import com.linkedin.venice.integration.utils.DaVinciTestContext;
 import com.linkedin.venice.integration.utils.ServiceFactory;
@@ -43,7 +39,6 @@ import com.linkedin.venice.integration.utils.VeniceTwoLayerMultiRegionMultiClust
 import com.linkedin.venice.meta.StoreInfo;
 import com.linkedin.venice.meta.Version;
 import com.linkedin.venice.meta.VersionStatus;
-import com.linkedin.venice.pushmonitor.ExecutionStatus;
 import com.linkedin.venice.utils.IntegrationTestPushUtils;
 import com.linkedin.venice.utils.RegionUtils;
 import com.linkedin.venice.utils.StoreMigrationTestUtil;
@@ -263,170 +258,6 @@ public class TestDeferredVersionSwap {
           false,
           -1);
       assertTrue(versionCreationResponse.isError());
-    }
-  }
-
-  @Test(timeOut = TEST_TIMEOUT * 2, dataProvider = "regionsProvider")
-  public void testDeferredVersionSwapWithFailedPushInNonTargetRegions(String failingRegions) throws IOException {
-    // Always mark the status as ERROR in regions
-    Set<String> failingRegionsList = RegionUtils.parseRegionsFilterList(failingRegions);
-    for (String failingRegion: failingRegionsList) {
-      multiRegionMultiClusterWrapper.failPushInRegion(failingRegion);
-    }
-
-    // Setup job properties
-    File inputDir = getTempDataDirectory();
-    TestWriteUtils.writeSimpleAvroFileWithStringToV3Schema(inputDir, 100, 100);
-    String inputDirPath = "file://" + inputDir.getAbsolutePath();
-    String storeName = Utils.getUniqueString("testDeferredVersionSwapWithFailedPushInNonTargetRegions");
-    Properties props =
-        IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, inputDirPath, storeName);
-    String keySchemaStr = "\"string\"";
-    UpdateStoreQueryParams storeParms = new UpdateStoreQueryParams().setUnusedSchemaDeletionEnabled(true);
-    storeParms.setTargetRegionSwapWaitTime(1);
-    String parentControllerURLs = multiRegionMultiClusterWrapper.getControllerConnectString();
-
-    try (ControllerClient parentControllerClient = new ControllerClient(CLUSTER_NAMES[0], parentControllerURLs)) {
-      createStoreForJob(CLUSTER_NAMES[0], keySchemaStr, NAME_RECORD_V3_SCHEMA.toString(), props, storeParms).close();
-
-      // Start push job with target region push enabled and check that it fails
-      props.put(TARGETED_REGION_PUSH_WITH_DEFERRED_SWAP, true);
-      try {
-        TestWriteUtils.runPushJob("Test push job", props);
-      } catch (Exception e) {
-        assertEquals(e.getClass(), VeniceException.class);
-      }
-
-      // Wait for job to fail
-      TestUtils.waitForNonDeterministicAssertion(1, TimeUnit.MINUTES, true, () -> {
-        JobStatusQueryResponse jobStatusQueryResponse = assertCommand(
-            parentControllerClient.queryOverallJobStatus(Version.composeKafkaTopic(storeName, 1), Optional.empty()));
-        ExecutionStatus executionStatus = ExecutionStatus.valueOf(jobStatusQueryResponse.getStatus());
-        assertEquals(executionStatus, ExecutionStatus.ERROR);
-      });
-
-      TestUtils.waitForNonDeterministicAssertion(1, TimeUnit.MINUTES, () -> {
-        Map<String, Integer> coloVersions =
-            parentControllerClient.getStore(storeName).getStore().getColoToCurrentVersions();
-
-        coloVersions.forEach((colo, version) -> {
-          if (!failingRegionsList.contains(colo)) {
-            Assert.assertEquals((int) version, 1);
-          } else {
-            Assert.assertEquals((int) version, 0);
-          }
-        });
-      });
-
-      TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, () -> {
-        StoreInfo parentStore = parentControllerClient.getStore(storeName).getStore();
-        Assert.assertEquals(parentStore.getVersion(1).get().getStatus(), VersionStatus.PARTIALLY_ONLINE);
-      });
-
-      // Check that child version status is marked as ONLINE if it didn't fail
-      for (VeniceMultiClusterWrapper childDatacenter: childDatacenters) {
-        ControllerClient childControllerClient =
-            new ControllerClient(CLUSTER_NAMES[0], childDatacenter.getControllerConnectString());
-        if (!failingRegionsList.contains(childDatacenter.getRegionName())) {
-          StoreResponse store = childControllerClient.getStore(storeName);
-          Optional<Version> version = store.getStore().getVersion(1);
-          assertNotNull(version);
-          assertEquals(version.get().getStatus(), VersionStatus.ONLINE);
-        }
-      }
-
-      // Verify that we can create a new version
-      VersionCreationResponse versionCreationResponse = parentControllerClient.requestTopicForWrites(
-          storeName,
-          1000,
-          Version.PushType.BATCH,
-          Version.guidBasedDummyPushId(),
-          true,
-          true,
-          false,
-          Optional.empty(),
-          Optional.empty(),
-          Optional.empty(),
-          false,
-          -1);
-      assertFalse(versionCreationResponse.isError());
-    }
-  }
-
-  @Test(timeOut = TEST_TIMEOUT * 2, dataProvider = "regionsProvider")
-  public void testDeferredVersionSwapWithFailedPushInTargetRegions(String failingRegions) throws IOException {
-    // Always mark the status as ERROR in regions
-    Set<String> failingRegionsList = RegionUtils.parseRegionsFilterList(failingRegions);
-    for (String failingRegion: failingRegionsList) {
-      multiRegionMultiClusterWrapper.failPushInRegion(failingRegion);
-    }
-
-    // Setup job properties
-    File inputDir = getTempDataDirectory();
-    TestWriteUtils.writeSimpleAvroFileWithStringToV3Schema(inputDir, 100, 100);
-    String inputDirPath = "file://" + inputDir.getAbsolutePath();
-    String storeName = Utils.getUniqueString("testDeferredVersionSwapWithFailedPushInTargetRegions");
-    Properties props =
-        IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, inputDirPath, storeName);
-    String keySchemaStr = "\"string\"";
-    UpdateStoreQueryParams storeParms = new UpdateStoreQueryParams().setUnusedSchemaDeletionEnabled(true);
-    storeParms.setTargetRegionSwapWaitTime(1);
-    String parentControllerURLs = multiRegionMultiClusterWrapper.getControllerConnectString();
-
-    try (ControllerClient parentControllerClient = new ControllerClient(CLUSTER_NAMES[0], parentControllerURLs)) {
-      createStoreForJob(CLUSTER_NAMES[0], keySchemaStr, NAME_RECORD_V3_SCHEMA.toString(), props, storeParms).close();
-
-      // Start push job with target region push enabled and check that it fails
-      String targetRegions = REGION1 + ", " + REGION2;
-      props.put(TARGETED_REGION_PUSH_WITH_DEFERRED_SWAP, true);
-      props.put(TARGETED_REGION_PUSH_LIST, targetRegions);
-      try {
-        TestWriteUtils.runPushJob("Test push job", props);
-      } catch (Exception e) {
-        assertEquals(e.getClass(), VeniceException.class);
-      }
-
-      // Wait for job to fail
-      TestUtils.waitForNonDeterministicAssertion(1, TimeUnit.MINUTES, true, () -> {
-        JobStatusQueryResponse jobStatusQueryResponse = assertCommand(
-            parentControllerClient.queryOverallJobStatus(Version.composeKafkaTopic(storeName, 1), Optional.empty()));
-        ExecutionStatus executionStatus = ExecutionStatus.valueOf(jobStatusQueryResponse.getStatus());
-        assertEquals(executionStatus, ExecutionStatus.ERROR);
-      });
-
-      TestUtils.waitForNonDeterministicAssertion(1, TimeUnit.MINUTES, () -> {
-        Map<String, Integer> coloVersions =
-            parentControllerClient.getStore(storeName).getStore().getColoToCurrentVersions();
-
-        coloVersions.forEach((colo, version) -> {
-          if (targetRegions.contains(colo) && !failingRegionsList.contains(colo)) {
-            Assert.assertEquals((int) version, 1);
-          } else {
-            Assert.assertEquals((int) version, 0);
-          }
-        });
-      });
-
-      TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, () -> {
-        StoreInfo parentStore = parentControllerClient.getStore(storeName).getStore();
-        Assert.assertEquals(parentStore.getVersion(1).get().getStatus(), VersionStatus.ERROR);
-      });
-
-      // Verify that we can create a new version
-      VersionCreationResponse versionCreationResponse = parentControllerClient.requestTopicForWrites(
-          storeName,
-          1000,
-          Version.PushType.BATCH,
-          Version.guidBasedDummyPushId(),
-          true,
-          true,
-          false,
-          Optional.empty(),
-          Optional.empty(),
-          Optional.empty(),
-          false,
-          -1);
-      assertFalse(versionCreationResponse.isError());
     }
   }
 
@@ -688,117 +519,6 @@ public class TestDeferredVersionSwap {
 
       client2.close();
     }
-  }
-
-  @Test(timeOut = TEST_TIMEOUT)
-  public void testDvcDelayedIngestionWithFailingPushInTargetRegion() throws Exception {
-    // Setup job properties
-    UpdateStoreQueryParams storeParms = new UpdateStoreQueryParams().setUnusedSchemaDeletionEnabled(true);
-    storeParms.setTargetRegionSwapWaitTime(1);
-    String parentControllerURLs = multiRegionMultiClusterWrapper.getControllerConnectString();
-    String keySchemaStr = "\"int\"";
-    String valueSchemaStr = "\"int\"";
-
-    // Create store + start a normal push
-    int keyCount = 100;
-    File inputDir = getTempDataDirectory();
-    TestWriteUtils.writeSimpleAvroFileWithIntToIntSchema(inputDir, keyCount);
-    String inputDirPath = "file://" + inputDir.getAbsolutePath();
-    String storeName = Utils.getUniqueString("testDvcDelayedIngestionWithTargetRegion");
-    Properties props =
-        IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, inputDirPath, storeName);
-    try (ControllerClient parentControllerClient = new ControllerClient(CLUSTER_NAMES[0], parentControllerURLs)) {
-      createStoreForJob(CLUSTER_NAMES[0], keySchemaStr, valueSchemaStr, props, storeParms).close();
-      TestWriteUtils.runPushJob("Test push job", props);
-      TestUtils.waitForNonDeterministicPushCompletion(
-          Version.composeKafkaTopic(storeName, 1),
-          parentControllerClient,
-          30,
-          TimeUnit.SECONDS);
-
-      // Version should only be swapped in all regions
-      TestUtils.waitForNonDeterministicAssertion(1, TimeUnit.MINUTES, () -> {
-        Map<String, Integer> coloVersions =
-            parentControllerClient.getStore(storeName).getStore().getColoToCurrentVersions();
-
-        coloVersions.forEach((colo, version) -> {
-          Assert.assertEquals((int) version, 1);
-        });
-      });
-    }
-
-    verifyThatPushStatusStoreIsOnline(storeName);
-    verifyThatMetaStoreIsOnline(storeName);
-
-    // Create dvc client in target region
-    List<VeniceMultiClusterWrapper> childDatacenters = multiRegionMultiClusterWrapper.getChildRegions();
-    VeniceClusterWrapper cluster1 = childDatacenters.get(0).getClusters().get(CLUSTER_NAMES[0]);
-    VeniceProperties backendConfig = DaVinciTestContext.getDaVinciPropertyBuilder(cluster1.getZk().getAddress())
-        .put(DATA_BASE_PATH, Utils.getTempDataDirectory().getAbsolutePath())
-        .put(LOCAL_REGION_NAME, REGION2)
-        .put(CLIENT_SYSTEM_STORE_REPOSITORY_REFRESH_INTERVAL_SECONDS, 1)
-        .build();
-    DaVinciClient<Object, Object> client1 =
-        ServiceFactory.getGenericAvroDaVinciClient(storeName, cluster1, new DaVinciConfig(), backendConfig);
-    client1.subscribeAll().get();
-
-    // Check that v1 is ingested
-    for (int i = 1; i <= keyCount; i++) {
-      assertNotNull(client1.get(i).get());
-    }
-
-    // Fail push in target region
-    multiRegionMultiClusterWrapper.failPushInRegion(REGION2);
-
-    // Do another push with target region enabled
-    int keyCount2 = 200;
-    File inputDir2 = getTempDataDirectory();
-    String inputDirPath2 = "file://" + inputDir2.getAbsolutePath();
-    TestWriteUtils.writeSimpleAvroFileWithIntToIntSchema(inputDir2, keyCount2);
-    Properties props2 =
-        IntegrationTestPushUtils.defaultVPJProps(multiRegionMultiClusterWrapper, inputDirPath2, storeName);
-    try (ControllerClient parentControllerClient = new ControllerClient(CLUSTER_NAMES[0], parentControllerURLs)) {
-      props2.put(TARGETED_REGION_PUSH_WITH_DEFERRED_SWAP, true);
-      props2.put(TARGETED_REGION_PUSH_LIST, REGION2);
-      try {
-        TestWriteUtils.runPushJob("Test push job", props2);
-      } catch (Exception e) {
-        assertEquals(e.getClass(), VeniceException.class);
-      }
-
-      // Version shouldn't be swapped in any region
-      TestUtils.waitForNonDeterministicAssertion(1, TimeUnit.MINUTES, () -> {
-        Map<String, Integer> coloVersions =
-            parentControllerClient.getStore(storeName).getStore().getColoToCurrentVersions();
-
-        coloVersions.forEach((colo, version) -> {
-          Assert.assertEquals((int) version, 1);
-        });
-      });
-
-      // Version status should be ERROR
-      TestUtils.waitForNonDeterministicAssertion(90, TimeUnit.SECONDS, () -> {
-        StoreInfo parentStore = parentControllerClient.getStore(storeName).getStore();
-        Assert.assertEquals(parentStore.getVersion(2).get().getStatus(), VersionStatus.ERROR);
-      });
-
-      // verify that dvc client did not ingest the version
-      TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, () -> {
-        for (int i = 101; i <= keyCount2; i++) {
-          assertNull(client1.get(i).get());
-        }
-      });
-
-      // Wait for push completion
-      TestUtils.waitForNonDeterministicAssertion(30, TimeUnit.SECONDS, () -> {
-        String kafkaTopicName = Version.composeKafkaTopic(storeName, 2);
-        JobStatusQueryResponse response =
-            parentControllerClient.queryOverallJobStatus(kafkaTopicName, Optional.empty());
-        assertEquals(response.getStatus(), ExecutionStatus.ERROR.toString());
-      });
-    }
-
-    client1.close();
   }
 
   private void verifyThatPushStatusStoreIsOnline(String storeName) {
